requirements.txt
Flask==3.1.2
flask-cors==6.0.1
transformers==4.56.1
vllm==0.10.1.1
accelerate==1.10.1

PyMuPDF==1.23.6
psycopg2-binary==2.9.9
flasgger>=0.9.5
